+++ src/components/companies/sectors/scopes/EmissionsSourcesAnlaysis.tsx
