+++ src/components/companies/sectors/scopes/KeyInsights.tsx
