+++ src/components/companies/sectors/scopes/ScopeCard.tsx
