+++ src/components/companies/sectors/scopes/ScopeCards.tsx
