+++ src/components/companies/sectors/scopes/ScopeModal.tsx
