+++ src/components/companies/sectors/scopes/ValueChainOverview.tsx
