+++ src/components/companies/sectors/trends/EmissionsTrendAnalysis.tsx
