+++ src/components/companies/sectors/trends/TrendCard.tsx
