+++ src/components/companies/sectors/trends/TrendCards.tsx
