+++ src/components/companies/sectors/trends/TrendCompanyList.tsx
